requirements.txt
djangorestframework==3.14.0
django-extensions==3.2.3
ipython==8.15.0
drf-yasg==1.21.7